# doc/openapi.yaml
openapi: "3.0.3"
info:
  title: "Inthub Backend API"
  version: "1.0.0"
  description: >
    この仕様書は、FastAPI によるバックエンド実装に対応する API を定義しています。
    各エンドポイントには、詳細な説明、想定されるエラー(400, 401, 404, 409, 500 など)も記載しています。
servers:
  - url: "https://inthub-fjfjehdsc4akamdg.japaneast-01.azurewebsites.net"
...
paths:
  /api/v1/auth/signup/callback:
    get:
      tags:
        - Auth
      summary: "ユーザー登録（サインアップコールバック）"
      description: >
        Azure AD B2C のサインアップ完了後に呼ばれるコールバックエンドポイントです。
        クエリパラメータに含まれるIDトークンを受け取り、ユーザー属性を抽出して返却します。
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
          description: "Azure AD B2C のサインアップ完了後に呼ばれるコールバックエンドポイント。クエリパラメータに含まれるIDトークン"
      responses:
        '200':
          description: "ユーザー登録に成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignUpResponse"
        '400':
          description: "入力不備または必須フィールドの不足"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "サーバ内部エラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/signin/callback:
    get:
      tags:
        - Auth
      summary: "ユーザー認証（サインインコールバック）"
      description: >
        Azure AD B2C のサインイン完了後に呼ばれるコールバックエンドポイントです。
        クエリパラメータに含まれるIDトークンを受け取り、ユーザー属性を抽出して返却します。
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
          description: "Azure AD B2C のサインイン完了後に呼ばれるコールバックエンドポイント。クエリパラメータに含まれるIDトークン"
      responses:
        '200':
          description: "サインインに成功し、アクセストークン等を返却"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignInResponse"
        '401':
          description: "認証失敗：無効な資格情報"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "サーバ内部エラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/chat/sessions:
    post:
      tags:
        - Chat
      summary: "チャットセッション作成"
      description: >
        認証済みユーザーにより新規チャットセッション（チャットルーム）を作成します。  
        Authorization ヘッダーに Bearer トークンを必須とします。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionCreateRequest"
      responses:
        '200':
          description: "チャットセッションの作成に成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSessionCreateResponse"
        '401':
          description: "認証失敗：無効または不足しているトークン"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "サーバ内部エラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      tags:
        - Chat
      summary: "チャットセッション一覧取得"
      description: >
        認証済みユーザーのチャットセッション一覧を取得します。  
        クエリパラメータ status（active, ended）によりフィルタリング可能です。
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, ended]
          description: "チャットセッションのステータス"
      responses:
        '200':
          description: "チャットセッション一覧の取得に成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSessionListResponse"
        '401':
          description: "認証失敗：無効または不足しているトークン"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "サーバ内部エラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/chat/messages:
    post:
      tags:
        - Chat
      summary: "チャットメッセージ送信"
      description: >
        指定されたチャットセッション内で、ユーザーのメッセージを送信します。  
        送信後、システムによる回答が返却されます。認証済みユーザーのみ利用可能です。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageSendRequest"
      responses:
        '200':
          description: "チャットメッセージ送信に成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessageSendResponse"
        '400':
          description: "リクエストパラメータが不正"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          description: "認証失敗：無効または不足しているトークン"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: "対象のチャットセッションが存在しない"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "サーバ内部エラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      tags:
        - Chat
      summary: "チャットメッセージ一覧取得"
      description: >
        指定したチャットセッション ID に紐づく全メッセージを取得します。  
        クエリパラメータ session_id を必須とし、認証済みユーザーのみアクセス可能です。
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: session_id
          required: true
          schema:
            type: string
          description: "チャットルームID"
      responses:
        '200':
          description: "チャットメッセージ一覧の取得に成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessageListResponse"
        '400':
          description: "無効な session_id が指定された場合"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          description: "認証失敗：無効または不足しているトークン"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: "チャットセッションが存在しない"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "サーバ内部エラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/chat/sessions/end:
    post:
      tags:
        - Chat
      summary: "チャットセッション終了"
      description: >
        指定したチャットセッションの状態を "ended" に更新し、セッションを終了します。  
        認証済みユーザーのみアクセス可能です。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionEndRequest"
      responses:
        '200':
          description: "チャットセッションの終了に成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSessionEndResponse"
        '400':
          description: "リクエストパラメータが不正"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          description: "認証失敗：無効または不足しているトークン"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: "対象のチャットセッションが存在しない"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "サーバ内部エラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SignUpResponse:
      type: object
      properties:
        name:
          type: string
        company_id:
          type: string
        email:
          type: string
        azure_user_id:
          type: string
        role:
          type: string
      required:
        - name
        - company_id
        - email
        - azure_user_id
        - role
    SignInResponse:
      type: object
      properties:
        name:
          type: string
        company_id:
          type: string
        access_token:
          type: string
        azure_user_id:
          type: string
        role:
          type: string
        refresh_token:
          type: string
      required:
        - name
        - company_id
        - access_token
        - azure_user_id
        - role
        - refresh_token
    ChatSessionCreateRequest:
      type: object
      properties:
        room_name:
          type: string
          nullable: true
    ChatSessionCreateResponse:
      type: object
      properties:
        status:
          type: string
        session_id:
          type: string
        message:
          type: string
      required:
        - status
        - session_id
        - message
    ChatMessageSendRequest:
      type: object
      properties:
        session_id:
          type: string
        user_message:
          type: string
        senpai_id:
          type: string
      required:
        - session_id
        - user_message
        - senpai_id
    ChatMessageSendResponse:
      type: object
      properties:
        status:
          type: string
        system_reply:
          type: string
        system_score:
          type: array
          items:
            type: number
      required:
        - status
        - system_reply
        - system_score
    ChatSessionEndRequest:
      type: object
      properties:
        session_id:
          type: string
      required:
        - session_id
    ChatSessionEndResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
      required:
        - status
        - message
    ChatSessionItem:
      type: object
      properties:
        session_id:
          type: string
        room_name:
          type: string
        status:
          type: string
          enum: [active, ended]
        created_at:
          type: string
          format: date-time
      required:
        - session_id
        - room_name
        - status
        - created_at
    ChatSessionListResponse:
      type: object
      properties:
        status:
          type: string
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/ChatSessionItem"
      required:
        - status
        - sessions
    ChatMessageItem:
      type: object
      properties:
        message_id:
          type: string
        content:
          type: string
        is_assistant:
          type: boolean
        senpai_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required:
        - message_id
        - content
        - is_assistant
        - created_at
    ChatMessageListResponse:
      type: object
      properties:
        status:
          type: string
        session_id:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessageItem"
      required:
        - status
        - session_id
        - messages
...
    ErrorResponse:
      type: object
      properties:
        error_code:
          type: string
        error_message:
          type: string
        details:
          type: object
          nullable: true
          additionalProperties: true
      required:
        - error_code
        - error_message